Create pl_copy_nyc_files_from_landing_till_stg.json
This pipeline will read all files from lakehouse one by one, and copy to staging table in warehouse, if that file was already not processed earlier. It will first copy from lakehouse to landing layer, then from landing to staging. While copying to staging, it will filter out records that are out of a particular date range. After the copy to staging is complete, metadata is stored in stage metadata table to indicate that the file has already been processed and landing table is cleared